Completed encryption/decryption state machine
State machine is written to be fully compliant with C
(even though Arduino supports C++). I wanted to test out
the state machine design pattern in Patterns in C and it worked out
really, really well! The program more or less worked on the first try,
aside from small changes.

Now uses a different input format: byte 0 contains the size of the
upcoming input, then the following N bytes contain the input, and the
last byte contains either 'E' or 'D' to select encryption or
decryption respectively.

The state machine is set up to transition on a time step as
well as on a character input. By default, states will do
nothing in response to either event. In their "constructors"
(the "transition_*" functions) they set up a static function pointer
as an event handler. To transition to another state, they call
that state's transition function. This allows the state struct whose
pointer gets passed around to change its behavior (in line with what
state it's in) at runtime.